[CI] Avoid GHA cache potentially prefix-matching indices unexpectedly
index.project.git-cinnabar.build.<sha>.linux.x86_64.coverage could
potentially be used instead of
index.project.git-cinnabar.build.<sha>.linux.x86_64 when the former task
ends before the latter starts.
That would lead to a build error because `cache-hit` is still false in
that case, and the build script would try to copy its artifact in place
of the cached one and fail to do so because it doesn't allow overwrites.